+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-### 1.1.7
+### 1.2.0
+- Now requires Dart 3.
 - `BoardController` can now take `overlays` and `underlays` as parameters, and passes them through to `Board`.
 - `BoardBuilder` now takes a parameter `forceSquareAlignment`, which defaults to true, and forces elements to be aligned to their squares (previously this was not the case).
 - `SquareBuilder` functions (such as those used by `BoardBuilder`) now return `Widget?`, i.e. they can return null. `BoardBuilder` will build an empty container in this case.